+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add JSON-Dry support for `Map` and `Set`
 * Add the `HashSet` class
 * Make `Array.cast()` return array versions of `Map` and `Set` instances
+* Fix `Blast.checksumSymbol` not always being available
 
 ## 0.7.19 (2022-06-19)